requirements.txt
Flask>=3.1.1
Werkzeug>=3.1.3
requests>=2.32.4
python-dotenv>=1.0.0
gunicorn>=23.0.0
